Add Canvas module tracking to CS221Bot
In addition, we fix a bug that caused the live_channels list
in canvas.json to get reset every time the bot starts.

Important: to get the module tracking set up, you need to
execute `!unlive` then `!live` in every channel that is currently
live-tracking a course. This only needs to be done once. It needs to
be done because the module tracking needs to know which channels
are using live tracking.