Make !track and !untrack update canvas.json correctly; fix a type hinting bug
Previously, !track and !untrack would only update the value associated
with the "courses" key. The values associated with the "due_week"
and "due_day" keys were not updated in the file, even though these
values may have changed inside the program. This commit fixes this issue.

I tried to be smart in my previous commit and added the type hint "Tuple[str]"
to the "*args"-type parameters used in Discord commands. That results in
"TypeError: Type Tuple cannot be instantiated", so I'm removing those type hints.